use eval context in growthbook
